address refactoring feedback
