+++ home/screens/AccountModal/index.tsx
@@ -1,20 +1,18 @@
-import { borderRadius, CheckIcon, iconSize, spacing, UsersIcon } from '@expo/styleguide-native';
+import { spacing } from '@expo/styleguide-native';
 import { Ionicons } from '@expo/vector-icons';
-import { Text, View, Image, useExpoTheme, Row, Spacer, Divider } from 'expo-dev-client-components';
+import { Text, View, useExpoTheme, Row, Spacer } from 'expo-dev-client-components';
 import React from 'react';
-import { ActivityIndicator, FlatList, Platform } from 'react-native';
-import { useDispatch, useSelector } from 'redux/Hooks';
-import SessionActions from 'redux/SessionActions';
+import { ActivityIndicator, Platform } from 'react-native';
+import { useSelector } from 'redux/Hooks';
 import isUserAuthenticated from 'utils/isUserAuthenticated';
 
 import { PressableOpacity } from '../../components/PressableOpacity';
 import { useHome_CurrentUserQuery } from '../../graphql/types';
-import { useAccountName } from '../../utils/AccountNameContext';
+import { LoggedInAccountView } from './LoggedInAccountView';
 import { LoggedOutAccountView } from './LoggedOutAccountView';
 import { ModalHeader } from './ModalHeader';
 
 export function AccountModal() {
-  const { accountName, setAccountName } = useAccountName();
   const theme = useExpoTheme();
 
   const { data, loading, error, refetch } = useHome_CurrentUserQuery();
@@ -25,11 +23,6 @@ export function AccountModal() {
       isAuthenticated,
     };
   });
-  const dispatch = useDispatch();
-
-  const onSignOutPress = React.useCallback(() => {
-    dispatch(SessionActions.signOut());
-  }, [dispatch]);
 
   if (isAuthenticated && !data?.viewer) refetch(); // get accounts info after logging in
 
@@ -104,76 +97,7 @@ export function AccountModal() {
     <View flex="1">
       {Platform.OS === 'ios' && <ModalHeader />}
       {data?.viewer?.accounts ? (
-        <View flex="1" padding="medium">
-          <View bg="default" border="hairline" overflow="hidden" rounded="large">
-            <FlatList<typeof data.viewer.accounts[number]>
-              data={data.viewer.accounts}
-              keyExtractor={(account) => account.id}
-              renderItem={({ item: account }) => (
-                <PressableOpacity
-                  key={account.id}
-                  containerProps={{ bg: 'default', style: { padding: 16 } }}
-                  onPress={() => {
-                    setAccountName(account.name);
-                  }}>
-                  <Row justify="between">
-                    <Row align={!account.owner?.fullName ? 'center' : 'start'}>
-                      {account?.owner?.profilePhoto ? (
-                        <Image
-                          size="xl"
-                          rounded="full"
-                          source={{ uri: account.owner.profilePhoto }}
-                        />
-                      ) : (
-                        <View rounded="full" height="xl" width="xl" bg="secondary" align="centered">
-                          <UsersIcon color={theme.icon.default} size={iconSize.small} />
-                        </View>
-                      )}
-                      <Spacer.Horizontal size="small" />
-                      <View>
-                        {account.owner ? (
-                          <>
-                            {account.owner.fullName ? (
-                              <>
-                                <Text type="InterBold">{account.owner.fullName}</Text>
-                                <Spacer.Vertical size="tiny" />
-                                <Text color="secondary" size="small">
-                                  {account.owner.username}
-                                </Text>
-                              </>
-                            ) : (
-                              <Text type="InterBold">{account.owner.username}</Text>
-                            )}
-                          </>
-                        ) : (
-                          <Text type="InterBold">{account.name}</Text>
-                        )}
-                      </View>
-                    </Row>
-                    {accountName === account.name && (
-                      <CheckIcon color={theme.icon.default} size={iconSize.large} />
-                    )}
-                  </Row>
-                </PressableOpacity>
-              )}
-              ItemSeparatorComponent={Divider}
-            />
-          </View>
-          <Spacer.Vertical size="large" />
-          <PressableOpacity
-            onPress={onSignOutPress}
-            style={{
-              backgroundColor: theme.button.tertiary.background,
-              padding: spacing[3],
-              justifyContent: 'center',
-              alignItems: 'center',
-            }}
-            borderRadius={borderRadius.medium}>
-            <Text style={{ color: theme.button.tertiary.foreground }} type="InterSemiBold">
-              Log Out
-            </Text>
-          </PressableOpacity>
-        </View>
+        <LoggedInAccountView accounts={data.viewer.accounts} />
       ) : (
         <LoggedOutAccountView />
       )}
